Silence some more warnings.

+++ rts/c/panic.h
@@ -20,7 +20,7 @@ static void panic(int eval, const char *fmt, ...)
 static char* msgprintf(const char *s, ...) {
   va_list vl;
   va_start(vl, s);
-  size_t needed = 1 + vsnprintf(NULL, 0, s, vl);
+  size_t needed = 1 + (size_t)vsnprintf(NULL, 0, s, vl);
   char *buffer = (char*) malloc(needed);
   va_start(vl, s); /* Must re-init. */
   vsnprintf(buffer, needed, s, vl);
